Add more callbacks and fix wrong implementation about freeing of data
The implementation of when to free the data in RushClient is not correct as
ngtcp2 is still holding on a reference to these pointers. According to ngtcp2
documentation, these data can only be freed after "being acked". Hence, this
diff fixes it by introducing more callbacks (some that will be used in the
future)

I have also added a unit test for RushClient to verify that the buffer
management is working as intended (when RushClient frees the data and keep
track of the buffers to send to ngtcp2).